Backend finished

+++ backend/src/main/java/com/pomidor/boil/transport/dtos/RecipientDto.java
@@ -1,8 +1,8 @@
 package com.pomidor.boil.transport.dtos;
 
 public record RecipientDto (
-    Long id,
-    Double demand,
+    Integer id,
+    Integer demand,
     Double unitPrice
 ) {
 }
