+++ backend/src/main/java/com/pomidor/boil/transport/dtos/SupplierDto.java
@@ -1,8 +1,8 @@
 package com.pomidor.boil.transport.dtos;
 
 public record SupplierDto(
-    Long id,
-    Double supply,
+    Integer id,
+    Integer supply,
     Double unitPrice
 ) {
 }
