+++ backend/src/main/java/com/pomidor/boil/transport/dtos/TransactionDto.java
@@ -1,8 +1,8 @@
 package com.pomidor.boil.transport.dtos;
 
 public record TransactionDto(
-    Long supplierId,
-    Long recipientId,
-    Double amount
+    Integer supplierId,
+    Integer recipientId,
+    Integer amount
 ) {
 }
