+++ backend/src/main/java/com/pomidor/boil/transport/dtos/TransportCostDto.java
@@ -1,8 +1,8 @@
 package com.pomidor.boil.transport.dtos;
 
 public record TransportCostDto(
-    Long supplierId,
-    Long recipientId,
+    Integer supplierId,
+    Integer recipientId,
     Double cost
 ) {
 }
